+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## []
+## [1.3.0] - 2023-11-08
 - Implement Slick reporting media override feature + docs
 - Add `Integrating reports into your Admin site` section to the docs
 - Group by and crosstab reports do not need date_field set anymore. Only time series do.